refactor: set layout as server component
